Bruh.

+++ main.py
@@ -8,12 +8,19 @@ from tqdm import tqdm
 from binascii import hexlify
 from imageio import mimsave as gif_save, imread as screen_read
 
+
+if not isfile('donatty.secret'):
+    print('Donatty token not found.')
+    exit(1)
+
+with open('donatty.secret') as file:
+    donatty = file.read()
+
 chrome_options = Options()
 chrome_options.add_argument("--headless")
 driver = webdriver.Chrome(options=chrome_options)
-url = "https://widgets.donatty.com/stats/?ref=d1a1acc8-e047-4a49-adf1-e1e3aed64d5e&token=6UoXLFXaM6zHfmw0v1vLO1sVwe5cmd"
-driver.get(url)
-driver.set_window_size(800, 800)
+driver.get(donatty)
+driver.set_window_size(800, 210)
 sleep(5)
 if isdir('screens'):
     rmtree('screens')
@@ -21,7 +28,7 @@ if isfile('output.gif'):
     remove('output.gif')
 mkdir('screens')
 chdir('screens')
-for i in tqdm(range(300)):
+for i in tqdm(range(400)):
     screen_id = hexlify(int.to_bytes(i, length=2, byteorder='big')).decode()
     driver.save_screenshot(f'screen-{screen_id}.png')
     sleep(1/30)
